catalog-service/cloudFunction/requirements.txt
functions-framework==3.8.3
mysql-connector-python==9.6.0
python-dotenv==1.2.2